+++ src/getStepStartStates.test.ts
@@ -80,7 +80,7 @@ import {
   STEP_RM_BATCH_APPLICATION,
   STEP_RM_BATCH_CERTIFICATE,
 } from './steps/resource-manager/batch';
-
+import { STEP_RM_REDIS_CACHES } from './steps/resource-manager/redis-cache';
 import {
   STEP_RM_EVENT_GRID_DOMAINS,
   STEP_RM_EVENT_GRID_DOMAIN_TOPICS,
@@ -154,6 +154,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_POOL]: { disabled: true },
       [STEP_RM_BATCH_APPLICATION]: { disabled: true },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
+      [STEP_RM_REDIS_CACHES]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -221,6 +222,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_POOL]: { disabled: true },
       [STEP_RM_BATCH_APPLICATION]: { disabled: true },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
+      [STEP_RM_REDIS_CACHES]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -288,6 +290,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_POOL]: { disabled: false },
       [STEP_RM_BATCH_APPLICATION]: { disabled: false },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: false },
+      [STEP_RM_REDIS_CACHES]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: false },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: false },